Begin coding different UI classes
